refactor: 패키지 이동 및 로또 생성 전략패턴 추가
랜덤 생성뿐만 아니라 수동전략과 같은 여러 로또 생성 전략을 확장할 수
있도록 interface를 만들어 로또 생성 전략패턴을 추가하였다.
