+++ src/main/java/lotto/LottoNumberGenerator.java
@@ -1,0 +1,8 @@
+package lotto;
+
+import java.util.List;
+import lotto.domain.Lotto;
+
+public interface LottoNumberGenerator {
+    List<Lotto> generate(int amount);
+}
